fix db

+++ db/migrate/20250328042038_create_checks.rb
@@ -1,7 +1,7 @@
 class CreateChecks < ActiveRecord::Migration[8.0]
   def change
     create_table :checks do |t|
-      t.integer :number, null: false
+      t.bigint :number, null: false
       t.references :company, null: false, foreign_key: true
 
       t.timestamps
